backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1  # For database migrations

# Authentication & Security
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
cryptography==41.0.8
python-multipart==0.0.6

# Browser Automation
playwright==1.40.0

# HTTP & Web Scraping
aiohttp==3.9.1
beautifulsoup4==4.12.2
requests==2.31.0

# Environment & Configuration
python-dotenv==1.0.0

# Utilities
asyncio-throttle==1.0.2
retrying==1.3.4

# Development & Testing (optional, install with pip install -r requirements-dev.txt)
# pytest==7.4.3
# pytest-asyncio==0.21.1
# pytest-cov==4.1.0
# black==23.12.0
# flake8==6.1.0
# mypy==1.7.1

# Production (optional, for deployment)
# gunicorn==21.2.0
# redis==5.0.1  # If using Redis for caching/queues
# celery==5.3.4  # If using task queues

# Monitoring (optional)
# prometheus-fastapi-instrumentator==6.1.0
# sentry-sdk==1.39.1